Add support for an arbitrary prefix
By default, dotfiles are stored with no prefix, so ~/.bashrc will link
to ~/Dotfiles/bashrc. Some folks, however, are already storing their
dotfiles in a repository where they have a prefix. '.' is common, but
I've seen one using '_' as well. This feature will allow those people to
specify their prefix in the configuration file without renaming all the
files in their repository. To do this, a configuration like:

    [dotfiles]
    prefix = .

in ~/.dotfilesrc is all that is needed.