Handle symlink creation on Windows
On Windows, os.symlink takes a third parameter to indicate whether the
symlink is to be a file or directory symlink.  This change creates a
wrapper function named symlink.  On non-Windows OSes, it is just an alias
to os.symlink.  On Windows, it passes the necessary third parameter to
os.symlink based on whether the source is a file or a directory (and thus
requires that the source actually exist).  All uses of os.symlink are
replaced with calls to the new wrapper function.