Add unit test for adding a missing package
This fails because the add operation is not aware of package settings,
and therefore adds a package dotfiles as it normally would - as
a toplevel symlink.  The problem arises when a user attempts to sync and
the expected symlink layout is not as it should be.  With the force
option, this can lead to deleted files with no way to recover.

Re #20, #27