Prevent adding toplevel package directories
This change prevents users from adding toplevel package directories via
the add operation in the cli.  Until add is package-aware, allowing this
behaviour can lead to loss of configuration files if the --force option
is later used to sync.

Adding packages already-defined packages will be possible in the future,
but this patch is necessary to prevent users from reaching an invalid
state.

Re #20, #27