Add dotfile class rework
This version of the dotfile class uses py.path.local and removes much of
the cruft of using os.path and it's associated requirements.

There are still a few things missing, namely dry_run mode where commands
are printed instead of executed and force sync.  They will be included
as required.