Add link scheme for flat repository
This commit also includes support for all of the CLI commands.  This is
for a "flat" repository where dotfiles must be proper files, no
directories or files in subdirectories.  Support for that is coming and
should cover what used to be called "packages" in a more automatic way
without requiring configuration file settings.

At the moment, this is the bare minimum of useful functionality: flat
repository, no 'ignores', no 'externals', no 'packages', and no
'prefixes'.  In this way I can make sure I get the object model and
abstractions correct before building on it further.  This I feel is what
the old version suffered from and caused difficulties in maintenance.

The above-mentioned features will return in some form as I move through
each of them, not to worry.  I'd like to see how far I can get without
any configuration file before any of that stuff gets added back in.