Add initial support for nesting
This feature means to replace what was called 'packages'.  The old
packages implementation was only one subdirectory lower than the home
directory and quite limiting in that regard.  Even worse, they had to be
configured manually in the configuration file and could not be created
from the command line.

This approach offers any depth the user specifies because the rule is:
you can only add files to the repository, not directories.  So if a user
adds a directory, the effect will be that each file contained in any
portion of that directory will be moved into the repository and a
symlink created.  This completely removes both manual configuration and
link ambiguity because there is always only once choice.  The repository
is used to maintain the directory structure, but only the leaf nodes
(files) are symlinked in the home directory.

This also makes it easy to store other non-dotfile files in the
repository like wallpapers, fonts, music, or whatever.