Add support for configuration files
At the moment, only a global configuration file is considered at
'~/.config/dotfiles/config.ini'.  Further more, only the 'repositories'
field is being used at the moment, but this will change in the next
commit.  If you to specify a repository (or several) in the config file,
you could use something like:

  [dotfiles]
  repositories = ['~/public/dotfiles', '~/private/dotfiles']

Still some polishing to be done, but it's coming along.