Begin adding support for file copy
Until now, dotfiles has only linked files using symbolic links.  This
doesn't always work, some configuration files cannot be symlinks.  This
commit begins adding support for file copy instead of linking.  File
status is still maintained, if the dotfile is copied and the content
differ from what's in the repository, a conflict is displayed.
Otherwise the file is an exact duplicate and that's what we want.

Still to do is exposing a command to copy from the cli and adding
support to remove() to do the right thing.  It may also be nice to see
file status (symlink vs. copy) in status output, still undecided.